requirements.txt
PyYAML==3.12
astropy==1.3.3
backports.functools-lru-cache==1.3
chainmap==1.0.2
elasticsearch==5.4.0
lxml==3.8.0
mail-parser==1.2.2
patool==1.12
pyparsing==2.2.0
python-magic==0.4.12
redis==2.10.5
shodan==1.7.1
simplejson==3.10.0
six==1.10.0
ssdeep==3.2
streamparse==3.6.0
tika-app==1.1.1
virustotal-api==1.1.2
